Add imuc_ctx and a graph implementation

+++ compiler/imuc_ir/src/sym/templ.rs
@@ -1,1 +1,1 @@
-pub type TemplId = u8;
+pub type TemplId = u32;
